file structure and module names
